+++ jdbc-postgres/src/main/java/io/kestra/repository/postgres/PostgresExecutionRepositoryService.java
@@ -1,0 +1,68 @@
+package io.kestra.repository.postgres;
+
+import io.kestra.core.models.QueryFilter;
+import io.kestra.core.models.executions.Execution;
+import io.kestra.core.utils.Either;
+import io.kestra.jdbc.AbstractJdbcRepository;
+import org.jooq.Condition;
+import org.jooq.impl.DSL;
+
+import java.util.*;
+
+import static io.kestra.core.models.QueryFilter.Op.EQUALS;
+
+public abstract class PostgresExecutionRepositoryService {
+    public static Condition findCondition(AbstractJdbcRepository<Execution> jdbcRepository, String query, Map<String, String> labels) {
+        List<Condition> conditions = new ArrayList<>();
+
+        if (query != null) {
+            conditions.add(jdbcRepository.fullTextCondition(Collections.singletonList("fulltext"), query));
+        }
+
+        if (labels != null)  {
+            labels.forEach((key, value) -> {
+                String sql = "value -> 'labels' @> '[{\"key\":\"" + key + "\", \"value\":\"" + value + "\"}]'";
+                conditions.add(DSL.condition(sql));
+            });
+        }
+
+        return conditions.isEmpty() ? DSL.trueCondition() : DSL.and(conditions);
+    }
+
+    public static Condition findLabelCondition(Either<Map<?, ?>, String> input, QueryFilter.Op operation) {
+        List<Condition> conditions = new ArrayList<>();
+        List<Condition> inConditions = new ArrayList<>();
+        if (input.isRight()) {
+            var query = input.right().get();
+            if (Objects.requireNonNull(operation) == QueryFilter.Op.CONTAINS) {
+                String sql = """
+                    EXISTS (
+                      SELECT 1
+                      FROM jsonb_array_elements(value -> 'labels') AS l
+                      WHERE l ->> 'value' ILIKE '%' || ? || '%'
+                    )
+                    """;
+                conditions.add(DSL.condition(sql, query));
+            } else {
+                throw new UnsupportedOperationException("Unsupported operation for query: " + operation);
+            }
+        } else {
+            var labels = input.getLeft();
+            labels.forEach((key, value) -> {
+                String sql = "value -> 'labels' @> '[{\"key\":\"" + key + "\", \"value\":\"" + value + "\"}]'";
+                switch (operation) {
+                    case EQUALS -> conditions.add(DSL.condition(sql));
+                    case NOT_EQUALS, NOT_IN -> conditions.add(DSL.not(DSL.condition(sql)));
+                    case IN -> inConditions.add(DSL.condition(sql));
+                    default -> throw new UnsupportedOperationException("Unsupported operation: " + operation);
+                }
+            });
+        }
+
+        if (!inConditions.isEmpty()) {
+            conditions.add(DSL.or(inConditions));
+        }
+        return conditions.isEmpty() ? DSL.trueCondition() : DSL.and(conditions);
+    }
+
+}
